fix Windows test failure in FixedDecimal.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/10273


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@34308 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/plurrule.cpp
@@ -1376,7 +1376,7 @@ FixedDecimal::FixedDecimal(const FixedDecimal &other) {
 
 
 void FixedDecimal::init(double n) {
-    int64_t numFractionDigits = decimals(n);
+    int32_t numFractionDigits = decimals(n);
     init(n, numFractionDigits, getFractionalDigits(n, numFractionDigits));
 }
